[Refactoring] int type to NSInteger

+++ LYCategory/Classes/_Foundation/_Date/NSDate+Formatter.h
@@ -51,7 +51,7 @@
 
 - (NSInteger)weekdayNumber;
 
-- (int)weekdayNumberPRC;
+- (NSInteger)weekdayNumberPRC;
 
 #pragma mark HOUR MINUTE SECOND
